requirements.txt
aiofiles==0.5.0
alembic==1.4.2
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.2.0
backpack==0.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cleo==0.6.8
click==7.1.2
confluent-kafka==1.5.0
ecdsa==0.14.1
Faker==0.8.18
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==2.0.0
inflect==4.1.0
inflection==0.3.0
iniconfig==1.1.1
Jinja2==2.11.2
lazy-object-proxy==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
orator==0.9.9
packaging==20.4
passlib==1.7.4
pastel==0.1.1
pathlib==1.0.1
pendulum==1.5.1
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyaml==16.12.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
pydantic-sqlalchemy==0.0.7
Pygments==2.7.2
pylev==1.3.0
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.19
starlette==0.13.6
text-unidecode==1.2
toml==0.10.1
tzlocal==1.5.1
urllib3==1.25.11
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
zipp==3.4.0
